#135139 - move setup class to src directory, revise coding to work that way, change workflow name
